# feat: Rework image generation to improve performance