--- a/.changeset/good-mirrors-bake.md
+++ b/.changeset/good-mirrors-bake.md
@@ -1,0 +1,5 @@
+---
+'astro': minor
+---
+
+Add concurrency to asset generation