## Changes

Change how image generation is done to improve performance:
- Images are now grouped under the base image (or URL) they originate from, as such, we can reuse the same buffer to generate all the variants, whereas before we would reload / refetch the image for every variant
- Adds a concurrent promise queue for image generations. Image generation plays well with being done concurrently, so it's a massive performance increase in a lot of cases, especially for uncached runs.
- Previously we'd do a lot of work to setup image generation on every image. This is now only done once and shared across all calls.

On my website, that has 222 local images, here's the result:

**Before**
- Fresh build: 26s~
- Cached build: 2s~

**After**
- Fresh build: 8s~
- Cached: 2s~

Not exactly scientific, but there's definitely an improvement in there.

## Testing

Tests should still pass!

## Docs

N/A.