@itsmatteomanf 2023-10-13
We need to be able to detect when a file is empty as with multiple writers/readers you can get into a situation in which the generation crashes with a file generated and not written.

This in turn causes the next build to fail with the `ENOENT` error. If the file is empty it should be ignored, imho.