@itsmatteomanf 2023-10-15
Do we want to consider a way to set the cpu count manually?

In some instances one might want to reduce the parallelisation if the backend doesn't support it for remote images, or something...