@Princesseuh 2023-10-15
@bluwy @ematipico I kinda rewrote it all from scratch 🙈. I updated the PR description to reflect what was done and would love to see a re-review from you both! (I made `os` lower case)