@Princesseuh 2023-10-16
> Can we add some tests with remote images? Especially with the expiration part. It would be great if we could log the user when an image is **not** used anymore in case it's expired.

We have tests with remote images, both cached and uncached ones! We also have tests for errors specific to remote images etc.