@Princesseuh 2023-10-18
> @Princesseuh not sure where the delays are, but the cached builds are definitely slower than before... the un-cached ones are incredibly fast in comparison, but the cached builds aren't that much faster than uncached now.

It's a mix of multiple things
- My comment here: https://github.com/withastro/astro/pull/8821#discussion_r1359954528 (especially for remote images)
- Concurrency doesn't necessarily play well with a lot of short operations like this

For local images on a fast hard drive, there's probably no big differences between this PR and main, but remote images suffer a bit more, I'll fix it