@itsmatteomanf 2023-10-18
> It's a mix of multiple things
>
> - My comment here: feat: Rework image generation to improve performance #8821 (comment) (especially for remote images)
> - Concurrency doesn't necessarily play well with a lot of short operations like this

It's somewhat baffling to me though, adding concurrency doesn't only not improve the performance (for what is basically the same process as the uncached images, just without the fetch, which is basically what has always been done and not much different bar the actual copying to what local images do), but slows it down by almost an order of magnitude... there must be something tiny somewhere that causes this.

> I'll fix it.

I have no doubts about that ahah